Move Web3SignerMetricCategory to common and inline the metric categories
