+++ common/src/main/java/tech/pegasys/web3signer/common/Web3SignerMetricCategory.java
@@ -10,18 +10,35 @@
  * an "AS IS" BASIS, WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied. See the License for the
  * specific language governing permissions and limitations under the License.
  */
-package tech.pegasys.web3signer.slashingprotection;
+package tech.pegasys.web3signer.common;
 
+import java.util.EnumSet;
 import java.util.Optional;
+import java.util.Set;
 
+import com.google.common.collect.ImmutableSet;
+import org.hyperledger.besu.metrics.StandardMetricCategory;
 import org.hyperledger.besu.plugin.services.metrics.MetricCategory;
 
-public enum SlashingMetricCategory implements MetricCategory {
-  ETH2_SLASHING_PROTECTION("eth2_slashingprotection");
+public enum Web3SignerMetricCategory implements MetricCategory {
+  HTTP("http"),
+  FILECOIN("filecoin"),
+  ETH2_SLASHING_PROTECTION("eth2_slashingprotection"),
+  SIGNING("signing");
 
   private final String name;
 
-  SlashingMetricCategory(final String name) {
+  public static final Set<MetricCategory> DEFAULT_METRIC_CATEGORIES;
+
+  static {
+    DEFAULT_METRIC_CATEGORIES =
+        ImmutableSet.<MetricCategory>builder()
+            .addAll(EnumSet.allOf(Web3SignerMetricCategory.class))
+            .addAll(EnumSet.allOf(StandardMetricCategory.class))
+            .build();
+  }
+
+  Web3SignerMetricCategory(final String name) {
     this.name = name;
   }
