+++ core/src/main/java/tech/pegasys/web3signer/core/service/http/metrics/HttpApiMetrics.java
@@ -12,9 +12,8 @@
  */
 package tech.pegasys.web3signer.core.service.http.metrics;
 
-import tech.pegasys.web3signer.core.metrics.Web3SignerMetricCategory;
+import tech.pegasys.web3signer.common.Web3SignerMetricCategory;
 import tech.pegasys.web3signer.signing.KeyType;
-import tech.pegasys.web3signer.signing.metrics.SigningMetricCategory;
 
 import org.hyperledger.besu.plugin.services.MetricsSystem;
 import org.hyperledger.besu.plugin.services.metrics.Counter;
@@ -36,12 +35,12 @@ public class HttpApiMetrics {
             "Number of requests received which had illegally formatted body.");
     signingTimer =
         metricsSystem.createTimer(
-            SigningMetricCategory.SIGNING,
+            Web3SignerMetricCategory.SIGNING,
             keyType.name().toLowerCase() + "_signing_duration",
             "Duration of a signing event");
     missingSignerCounter =
         metricsSystem.createCounter(
-            SigningMetricCategory.SIGNING,
+            Web3SignerMetricCategory.SIGNING,
             keyType.name().toLowerCase() + "_missing_identifier_count",
             "Number of signing operations requested, for keys which are not available");
   }
@@ -62,7 +61,7 @@ public class HttpApiMetrics {
     if (signersLoadedCounter == null) {
       signersLoadedCounter =
           metricsSystem.createCounter(
-              SigningMetricCategory.SIGNING,
+              Web3SignerMetricCategory.SIGNING,
               "signers_loaded_count",
               "Number of keys loaded (combining SECP256k1 and BLS12-381");
     }
