+++ signing/build.gradle
@@ -15,6 +15,8 @@ jar {
 }
 
 dependencies {
+  implementation project(":common")
+
   implementation 'com.fasterxml.jackson.dataformat:jackson-dataformat-yaml'
   implementation 'com.github.arteam:simple-json-rpc-server'
   implementation 'commons-io:commons-io'
