--- signing/src/main/java/tech/pegasys/web3signer/signing/metrics/FilecoinMetricCategory.java
@@ -1,37 +1,0 @@
-/*
- * Copyright 2022 ConsenSys AG.
- *
- * Licensed under the Apache License, Version 2.0 (the "License"); you may not use this file except in compliance with
- * the License. You may obtain a copy of the License at
- *
- * http://www.apache.org/licenses/LICENSE-2.0
- *
- * Unless required by applicable law or agreed to in writing, software distributed under the License is distributed on
- * an "AS IS" BASIS, WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied. See the License for the
- * specific language governing permissions and limitations under the License.
- */
-package tech.pegasys.web3signer.signing.metrics;
-
-import java.util.Optional;
-
-import org.hyperledger.besu.plugin.services.metrics.MetricCategory;
-
-public enum FilecoinMetricCategory implements MetricCategory {
-  FILECOIN("filecoin");
-
-  private final String name;
-
-  FilecoinMetricCategory(final String name) {
-    this.name = name;
-  }
-
-  @Override
-  public String getName() {
-    return name;
-  }
-
-  @Override
-  public Optional<String> getApplicationPrefix() {
-    return Optional.empty();
-  }
-}
